derive updated schemas from pydantic (GDEV-1320)
Define the schemas using pydantic BaseModels instead of JSON schemas.
JSON schemas can be autogenerated from these models (to be implemented
in another PR).
The schemas are now organized a schema registry
that is a key value store, where the keys correspond to
event types and the values to the event schemas defined
as BaseModels.

Moreover, the schema contents where modified to accomodate for changes
in the domain logic and to remove unnecessary fields.

Bump version to 0.5.0